Move transaction code into Core folder
