+++ src/Core/transactions/transaction-state.ts
@@ -1,0 +1,10 @@
+/** Transaction state. */
+export enum TransactionState {
+  Accepted = "Accepted",
+  Calling = "Calling",
+  Completed = "Completed",
+  Confirmed = "Confirmed",
+  Proceeding = "Proceeding",
+  Terminated = "Terminated",
+  Trying = "Trying"
+}
