+++ src/Core/user-agents/bye-user-agent-server.ts
@@ -1,7 +1,7 @@
 import { IncomingRequest as IncomingRequestMessage } from "../../SIPMessage";
-import { NonInviteServerTransaction } from "../../Transactions";
 import { InviteDialog } from "../dialogs";
 import { IncomingByeRequest, IncomingRequestDelegate } from "../messages";
+import { NonInviteServerTransaction } from "../transactions";
 import { UserAgentServer } from "./user-agent-server";
 
 export class ByeUserAgentServer extends UserAgentServer implements IncomingByeRequest {
