+++ src/Core/user-agents/info-user-agent-client.ts
@@ -1,7 +1,7 @@
 import { C } from "../../Constants";
-import { NonInviteClientTransaction } from "../../Transactions";
 import { InviteDialog } from "../dialogs";
 import { OutgoingInfoRequest, OutgoingRequestDelegate, RequestOptions } from "../messages";
+import { NonInviteClientTransaction } from "../transactions";
 import { UserAgentClient } from "./user-agent-client";
 
 export class InfoUserAgentClient extends UserAgentClient implements OutgoingInfoRequest {
