+++ src/Core/user-agents/info-user-agent-server.ts
@@ -1,7 +1,7 @@
 import { IncomingRequest as IncomingRequestMessage } from "../../SIPMessage";
-import { NonInviteServerTransaction,  } from "../../Transactions";
 import { InviteDialog } from "../dialogs";
 import { IncomingInfoRequest, IncomingRequestDelegate } from "../messages";
+import { NonInviteServerTransaction,  } from "../transactions";
 import { UserAgentServer } from "./user-agent-server";
 
 export class InfoUserAgentServer extends UserAgentServer implements IncomingInfoRequest {
