+++ src/Core/user-agents/publish-user-agent-client.ts
@@ -1,6 +1,6 @@
 import { OutgoingRequest as OutgoingRequestMessage } from "../../SIPMessage";
-import { NonInviteClientTransaction } from "../../Transactions";
 import { OutgoingPublishRequest, OutgoingRequestDelegate } from "../messages";
+import { NonInviteClientTransaction } from "../transactions";
 import { UserAgentCore } from "../user-agent-core";
 import { UserAgentClient } from "./user-agent-client";
